optional_plugins/html: move html test into plugin dir
