+++ selftests/functional/test_sysinfo.py
@@ -65,24 +65,6 @@ class SysInfoTest(TestCaseTmpDir):
                                                              result)
         self.assertFalse(os.path.isdir(sysinfo_dir), msg)
 
-    def test_sysinfo_html_output(self):
-        html_output = "{}/output.html".format(self.tmpdir.name)
-        cmd_line = ('{} run --html {} --job-results-dir {} '
-                    'passtest.py'.format(AVOCADO, html_output,
-                                         self.tmpdir.name))
-        result = process.run(cmd_line)
-        expected_rc = exit_codes.AVOCADO_ALL_OK
-        self.assertEqual(result.exit_status, expected_rc,
-                         'Avocado did not return rc %d:\n%s' % (expected_rc,
-                                                                result))
-        with open(html_output, 'rt') as fp:
-            output = fp.read()
-
-        # Try to find some strings on HTML
-        self.assertNotEqual(output.find('Filesystem'), -1)
-        self.assertNotEqual(output.find('root='), -1)
-        self.assertNotEqual(output.find('MemAvailable'), -1)
-
     def run_sysinfo_interrupted(self, sleep, timeout, exp_duration):
         commands_path = os.path.join(self.tmpdir.name, "commands")
         script.make_script(commands_path, "sleep %s" % sleep)
